+++ src/java/facade/PersonInterface.java
@@ -14,7 +14,7 @@ public interface PersonInterface {
     public Person createPerson(Person p);
     public Person deletePerson(Person p);
     public Person updatePerson(Person p);
-    public Person getPerson (int id);
+    public Person getPerson (long id);
     public List<Person> getPersons ();
  
 }
